index.html
Avail is a non-profit organization which offers free tutoring to kids grades 1-8. Older students can teach the younger ones via zoom, in order to recevie voulenteer hours. Retired teachers can also sign up in order to give back to the community. According to the oxford dictionary avail means to help or benefit, which is exactly what we are trying to do.
Our Purpose